finish controllers and services to send notification via rabbitmq and webhook

+++ carrito-service/src/main/java/com/arka/carrito_service/domain/useCases/EliminarDetalleUseCase.java
@@ -2,6 +2,7 @@ package com.arka.carrito_service.domain.useCases;
 
 import com.arka.carrito_service.domain.exception.DetalleCarritoNoEncontradoException;
 import com.arka.carrito_service.domain.model.Carrito;
+import com.arka.carrito_service.domain.model.Estado;
 import com.arka.carrito_service.domain.model.gateway.CarritoGateway;
 import com.arka.carrito_service.domain.model.gateway.DetalleCarritoGateway;
 import reactor.core.publisher.Mono;
@@ -16,16 +17,22 @@ public class EliminarDetalleUseCase {
         this.carritoGateway = carritoGateway;
     }
 
-    public Mono<Carrito>deleteDetalle(Integer idDetalleCarrito){
+    public Mono<Carrito>eliminarDetalle(Integer idDetalleCarrito){
         return detalleCarritoGateway.findById(idDetalleCarrito)
                 .switchIfEmpty(Mono.error(new DetalleCarritoNoEncontradoException(
                         "Car detail not found: " + idDetalleCarrito
                 )))
-                .flatMap(detalle ->
-                        detalleCarritoGateway.deleteById(idDetalleCarrito)
-                                .thenReturn(detalle.getIdCarrito())
-                )
-                .flatMap(idCarrito -> carritoGateway.findById(idCarrito));
+                .flatMap(detalle -> carritoGateway.findById(detalle.getIdCarrito())
+                        .flatMap(carrito -> {
+                            if (carrito.getEstado() != Estado.abierto) {
+                                return Mono.error(new IllegalStateException(
+                                        "No se puede modificar un carrito con estado: " + carrito.getEstado()
+                                ));
+                            }
+                            return detalleCarritoGateway.deleteById(idDetalleCarrito)
+                                    .thenReturn(detalle.getIdCarrito());
+                        })
+                .flatMap(idCarrito -> carritoGateway.findById(idCarrito)));
 
     }
 }
